pythonBackend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
djangorestframework_simplejwt==5.3.1
matplotlib==3.8.2
